Update backup_files.py
Removed report on directory creation because it was taking too much time without being too informative. Code is commented out with "# Uncomment to get report on directory " so it is easy to go back to if needed.
Also added reporting print statements on number of files copied and reviewed so user can see that things are happening.